extract exception formatting to utils
